requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
openai==1.0.0
requests==2.31.0
Pydantic==2.4.2
Pydantic[email]==2.4.2
gunicorn==21.2.0
python-dateutil==2.8.2
markdown==3.5.1
reportlab==4.0.7
Pillow==10.0.1
pytest==7.4.3
pytest-cov==4.1.0
beautifulsoup4==4.12.2
lxml==4.9.3
unidecode==1.3.0
